+++ src/sleep.js
@@ -2,8 +2,6 @@ var _ = require("./_");
 
 module.exports = function (time, val) {
     return new _.Promise(function (r) {
-        return setTimeout((function () {
-            return r(val);
-        }), time);
+        setTimeout(r, time, val);
     });
 };